+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to LTX Video Generator will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2025-01-09
+
+### Fixed
+- Fixed crash when viewing videos in history (AVKit SwiftUI crash on macOS 26)
+- Replaced SwiftUI VideoPlayer with NSViewRepresentable AVPlayerView wrapper
+
 ## [1.0.2] - 2025-01-09
 
 ### Fixed